+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add `RESTConfig()` for use in integration tests.
+
 ## [0.11.0] - 2021-06-16
 
 ### Added